Improve etc/profile handling

+++ shell/posix/functions/config.sh
@@ -240,13 +240,15 @@ _koopa_update_lmod_config() {                                             # {{{1
 _koopa_update_profile() {                                                 # {{{1
     # """
     # Link shared 'zzz-koopa.sh' configuration file into '/etc/profile.d/'.
-    # Updated 2019-11-05.
+    # Updated 2020-01-11.
     # """
-    local symlink
+    _koopa_is_shared || return 0
     _koopa_is_linux || return 0
     _koopa_has_sudo || return 0
+    local symlink
     symlink="/etc/profile.d/zzz-koopa.sh"
-    # > [ -L "$symlink" ] && return 0
+    # Early return if link already exists.
+    [ -L "$symlink" ] && return 0
     _koopa_message "Adding '${symlink}'."
     sudo rm -fv "/etc/profile.d/koopa.sh"
     sudo ln -fnsv \
